fix(backend): fetch only required attributes for group members
